Moved three classes to new file `database/messages.py`
