+++ website/api/socketio_handlers.py
@@ -5,8 +5,10 @@ This code contains the main functions that communicate with the server (server s
 from flask import request
 from flask_login import current_user
 
+from ..database.messages import Chat
+
 from ..database.player import Player
-from ..database.database import Chat, Message
+from ..database.messages import Message
 from ..utils import check_existing_chats
 from .. import db
 from datetime import datetime
