+++ website/views.py
@@ -6,9 +6,11 @@ from flask import Blueprint, render_template, request, flash
 from flask import g, current_app
 from flask_login import login_required, current_user
 
+from .database.messages import Chat
+
 from .database.player import Player
 from . import db
-from .database.database import Chat, Resource_on_sale
+from .database.database import Resource_on_sale
 from .utils import check_existing_chats
 
 views = Blueprint("views", __name__)
